+++ gr-blocks/lib/message_debug_impl.h
@@ -71,8 +71,8 @@ public:
     message_debug_impl(bool en_uvec);
     ~message_debug_impl() override;
 
-    int num_messages() override;
-    pmt::pmt_t get_message(int i) override;
+    size_t num_messages() override;
+    pmt::pmt_t get_message(size_t i) override;
     void set_vector_print(bool en) override { d_en_uvec = en; };
 };
